requirements.txt
# ============================================
# AI Recruitment Agent - Resume Parser + RAG
# Optimized for GTX 1650 Ti (4GB) / RTX 3050 (6GB)
# ============================================

# --- Core Framework ---
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.18

# --- Resume Parsing ---
pdfplumber==0.11.4
python-docx==1.1.2

# --- Embeddings (GPU-optimized, small footprint) ---
# all-MiniLM-L6-v2: ~80MB model, ~400MB VRAM
sentence-transformers==3.3.1
torch>=2.9.0

# --- Vector Store (Local, no server needed) ---
chromadb==0.5.23

# --- Text Processing ---
tiktoken==0.8.0

# --- Utilities ---
pydantic==2.10.3
pydantic-settings==2.7.0
python-dotenv==1.0.1
rich==13.9.4
jinja2==3.1.4
aiofiles==24.1.0
loguru==0.7.3
